Remove Omniauth Callbacks from Jumpstart namespace
Since Jumpstart is an isolated namespace, routes and other features
don't work as expected. If you use `root_path` for example, it will be
the engine's root, not the app's root.

This is unexpected, so I've refactored this to a concern instead. This
provides the same functionality but now works how you would expect.