Need to select a list before creating a member (#228)
* Need to select a list before creating a member

I was having an experience after trying to have new users added to my Mailchimp account where an error from Mailchimp was being returned about a resource not existing. I believe this was because the list was not specified. Per [gibbon documentation](https://github.com/amro/gibbon#subscribers), one would specify the list as shown in this change.

* Update comment with example

[Specifies a list](https://github.com/amro/gibbon#subscribers) for member to be added to so as to avoid error from Mailchimp API